Move to algorithms and parallelize Groth16 also
