+++ algorithms/Cargo.toml
@@ -147,6 +147,7 @@ default-features = false
 
 [dependencies.rayon]
 version = "1"
+optional = true
 
 [dependencies.rust-gpu-tools]
 version = "0.5.0"
@@ -240,6 +241,6 @@ msm = [ ]
 prf = [ ]
 signature = [ "encryption", "crypto_hash" ]
 snark = [ "fft", "msm" ]
-parallel = [ ]
+parallel = [ "rayon" ]
 print-trace = [ "snarkvm-profiler/print-trace" ]
 cuda = [ "rust-gpu-tools" ]
